requirements.txt
# -------------------------------
# ML / DL Core Libraries
# -------------------------------
numpy==1.26.4                  # Fundamental package for numerical computations
scipy==1.11.4                  # Scientific computing tools (linear algebra, etc.)
pandas==2.2.3                  # Data manipulation and analysis
scikit-learn==1.3.2            # Classical machine learning algorithms and tools
matplotlib==3.8.2              # Plotting and visualization library
seaborn==0.13.2                # Statistical data visualization
missingno==0.5.2               # Missing data visualization utilities
opencv-python-headless==4.9.0.80  # OpenCV without GUI support (image processing)
Pillow==10.3.0                 # Python Imaging Library (PIL) fork for image processing

# -------------------------------
# API Framework
# -------------------------------
fastapi==0.110.0               # High-performance web framework for building APIs
uvicorn[standard]==0.29.0      # ASGI server for running FastAPI
python-multipart==0.0.9        # Form data parsing (file uploads)

# -------------------------------
# Deep Learning - PyTorch (CUDA 12.4 Support)
# -------------------------------
torch @ https://download.pytorch.org/whl/cu124/torch-2.4.1%2Bcu124-cp310-cp310-linux_x86_64.whl  # PyTorch with CUDA 12.4
torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.19.1%2Bcpu-cp310-cp310-linux_x86_64.whl  # TorchVision CPU version

# -------------------------------
# Workflow Management & Model Serving
# -------------------------------
mlflow==2.9.2                  # Machine learning lifecycle management platform
tqdm==4.66.4                   # Progress bar utility for loops and data loading    
# hydra-core==1.3.2             # Config management (optional, commented)
# dvc[s3]==3.59.0               # Data version control with S3 (optional, commented)

# -------------------------------
# Object Storage Clients
# -------------------------------
minio==7.2.2                   # S3-compatible object storage client
boto3==1.34.0                  # AWS SDK for Python (S3, etc.)
